SDIT-1436 Reduce Adjudication listener threads to 8 (#473)
Connection pool size is 10 and default thread listener is 10 - when any other endpoint is called that requires a connection (e.g. healthcheck), it can use more connections than in the pool - hence 500 errors. This is a particular issue during early stages of Adjudication migration due to the get IDs endpoint being so CPU intensive.