SDIT-1519 Implement alert synchronization logic (#501)
This commit introduces alert synchronization logic in the service. If an ALERT-UPDATED event is received for an alert that has already been synchronized (checked by the audit module name), the event is ignored and a telemetry event is tracked. If the event is for a new alert, depending on whether the mapping exists or not, a failure or success telemetry event is tracked and appropriate actions are taken.