SDIT-2089 Correct slow complete migration logic
Previously migrations would incorrectly wait for the same amount time it would wait at the beginning of a migration for the initial slow "get ids" query compared to  the end of the migration.

This is now a separate config.

For contacts,allocations,activities this is set to 10 seconds all others are whatever the value is now to not break anything